moved enum TabId to domain/mainMenu.js
