+++ src/modules/menu/components/mainMenu/MainMenu.js
@@ -7,7 +7,8 @@ import { $injector } from '../../../../injection';
 import { DevInfo } from '../../../utils/components/devInfo/DevInfo';
 import { TopicsContentPanel } from '../../../topics/components/menu/TopicsContentPanel';
 import { SearchResultsPanel } from '../../../search/components/menu/SearchResultsPanel';
-import { TabId, toggle } from '../../../../store/mainMenu/mainMenu.action';
+import { toggle } from '../../../../store/mainMenu/mainMenu.action';
+import { TabId } from '../../../../domain/mainMenu';
 import { FeatureInfoPanel } from '../../../featureInfo/components/featureInfoPanel/FeatureInfoPanel';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 import { MapsContentPanel } from './content/maps/MapsContentPanel';
