+++ src/modules/search/components/menu/types/geoResource/GeoResourceResultItem.js
@@ -4,7 +4,8 @@
 import { html, nothing } from 'lit-html';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 import { addLayer, removeLayer } from '../../../../../../store/layers/layers.action';
-import { close as closeMainMenu, setTab, TabId } from '../../../../../../store/mainMenu/mainMenu.action';
+import { close as closeMainMenu, setTab } from '../../../../../../store/mainMenu/mainMenu.action';
+import { TabId } from '../../../../../../domain/mainMenu';
 import css from './geoResourceResultItem.css';
 import { MvuElement } from '../../../../../MvuElement';
 import { $injector } from '../../../../../../injection';
