+++ test/modules/search/components/menu/types/geoResource/GeoResourceResultItem.test.js
@@ -7,7 +7,7 @@ import {
 import { GeoResourceSearchResult } from '../../../../../../../src/modules/search/services/domain/searchResult';
 import { TestUtils } from '../../../../../../test-utils.js';
 import { createNoInitialStateMediaReducer } from '../../../../../../../src/store/media/media.reducer';
-import { TabId } from '../../../../../../../src/store/mainMenu/mainMenu.action';
+import { TabId } from '../../../../../../../src/domain/mainMenu';
 import { $injector } from '../../../../../../../src/injection';
 import { positionReducer } from '../../../../../../../src/store/position/position.reducer';
 import { Spinner } from '../../../../../../../src/modules/commons/components/spinner/Spinner';
