+++ test/plugins/ImportPlugin.test.js
@@ -9,7 +9,7 @@ import { layersReducer } from '../../src/store/layers/layers.reducer';
 import { LevelTypes } from '../../src/store/notifications/notifications.action';
 import { SourceType, SourceTypeName } from '../../src/domain/sourceType';
 import { createNoInitialStateMainMenuReducer } from '../../src/store/mainMenu/mainMenu.reducer';
-import { TabId } from '../../src/store/mainMenu/mainMenu.action';
+import { TabId } from '../../src/domain/mainMenu';
 import { positionReducer } from '../../src/store/position/position.reducer';
 
 describe('LAYER_ADDING_DELAY_MS', () => {
